Split out watch management into separate class
