+++ packages/yarnpkg-pnpify/sources/WatchManager.ts
@@ -1,0 +1,77 @@
+import {Filename, PortablePath, Watcher, WatchCallback} from '@yarnpkg/fslib';
+
+import {EventEmitter}                                   from 'events';
+
+import {NodePathResolver}                               from './NodePathResolver';
+
+class WatchEventEmitter extends EventEmitter {
+  private dirWatchers: DirectoryWatcherMap;
+  private watchPath: PortablePath;
+  private watcherId: number;
+
+  public constructor(dirWatchers: DirectoryWatcherMap, watchPath: PortablePath, watcherId: number) {
+    super();
+    this.dirWatchers = dirWatchers;
+    this.watchPath = watchPath;
+    this.watcherId = watcherId;
+  }
+
+  public close() {
+    const dirWatcher = this.dirWatchers.get(this.watchPath)!;
+    dirWatcher.eventEmitters.delete(this.watcherId);
+    if (Object.keys(dirWatcher.eventEmitters).length === 0) {
+      this.dirWatchers.delete(this.watchPath);
+    }
+  }
+}
+
+type DirectoryWatcherMap = Map<PortablePath, DirectoryWatcher>;
+
+interface DirectoryWatcher {
+  lastWatcherId: number;
+  eventEmitters: Map<number, Watcher & EventEmitter>;
+  dirEntries: Set<Filename>;
+}
+
+export class WatchManager extends EventEmitter {
+  private readonly dirWatchers: DirectoryWatcherMap = new Map();
+
+  public addWatcher(watchPath: PortablePath, dirList: Set<Filename>, callback: WatchCallback) {
+    let dirWatcher = this.dirWatchers.get(watchPath);
+    if (!dirWatcher) {
+      dirWatcher = {lastWatcherId: 0, eventEmitters: new Map(), dirEntries: dirList};
+      this.dirWatchers.set(watchPath, dirWatcher);
+    }
+    const watcherId = dirWatcher.lastWatcherId++;
+
+    const watchEventEmitter = new WatchEventEmitter(this.dirWatchers, watchPath, watcherId);
+    dirWatcher.eventEmitters.set(watcherId, watchEventEmitter);
+
+    watchEventEmitter.on('rename', (filename: string) => callback('rename', filename));
+  }
+
+  public notifyWatchers(pathResolver: NodePathResolver) {
+    for (const [watchPath, dirWatcher] of this.dirWatchers) {
+      const newDirEntries = pathResolver.resolvePath(watchPath).dirList || new Set();
+      // Difference between new and old directory contents
+      const dirEntryDiff = new Set();
+      for (const entry of newDirEntries) {
+        if (!dirWatcher.dirEntries.has(entry)) {
+          dirEntryDiff.add(entry);
+        }
+      }
+      for (const entry of dirWatcher.dirEntries) {
+        if (!newDirEntries.has(entry)) {
+          dirEntryDiff.add(entry);
+        }
+      }
+
+      for (const entry of dirEntryDiff) {
+        for (const watcher of Object.values(dirWatcher.eventEmitters)) {
+          watcher.emit('rename', entry);
+        }
+      }
+      dirWatcher.dirEntries = newDirEntries;
+    }
+  }
+}
